Allow missing dates in LAA Reference JSON response
According to the HMCTS schema specification for LAA Reference entity,
effectiveStartDate and effectiveEndDate are optional values.

During testing against this mock api was discovered that the Json
builder was blowing up as the Laa Reference stored in DB does not
contain those date values, but the Json builder assumes its existence
and tries to cast them into a date.

This commit fixes it by returning null for those values when they are
not defined in the database.